ultimo de clase

+++ lab2_T1G12/3_bugs/bug2.c
@@ -11,7 +11,7 @@
 
 int main (int argc, char *argv[]) {
     int numtasks, rank, tag=1, alpha, i;
-    float beta;
+    int beta;
     MPI_Request reqs[10];
     MPI_Status stats[10];
 
@@ -32,9 +32,9 @@ int main (int argc, char *argv[]) {
 
     if (rank == 1) {
         for (i=0; i<10; i++) {
-            MPI_Irecv(&beta, 1, MPI_FLOAT, 0, tag, MPI_COMM_WORLD, &reqs[i]);
+            MPI_Irecv(&beta, 1, MPI_INT, 0, tag, MPI_COMM_WORLD, &reqs[i]);
             MPI_Wait(&reqs[i], &stats[i]);
-            printf("Task %d received = %f\n",rank,beta);
+            printf("Task %d received = %d\n",rank,beta);
         }
     }
